+++ core/src/Settings.cpp
@@ -34,7 +34,7 @@ namespace Core::Service {
         m_localMusicPaths = json["MusicDirectory"].toVariant().toStringList();
         m_dbPath = json["DatabaseDirectory"].toString();
         m_userMusicList = json["UserLists"].toVariant().toStringList();
-        m_volume = static_cast<float>(json["DefaultVolume"].toDouble());
+        m_volume = json["DefaultVolume"].toInt();
     };
